refacto(categories): reorg files and directories
