+++ config/routes/Centreon/monitoring/category.yaml
@@ -3,11 +3,11 @@ centreon_application_monitoring_services_find_categories:
     methods: "GET"
     path: "/monitoring/services/categories"
     controller: >-
-        Core\Infrastructure\RealTime\Api\FindServiceCategory\FindServiceCategoryController
-    condition: "request.attributes.get('version') >= 22.04"
+        Core\Category\RealTime\Infrastructure\Api\FindServiceCategory\FindServiceCategoryController
+    condition: "request.attributes.get('version') >= 22.10"
 centreon_application_monitoring_host_find_categories:
     methods: "GET"
     path: "/monitoring/hosts/categories"
     controller: >-
-        Core\Infrastructure\RealTime\Api\FindHostCategory\FindHostCategoryController
-    condition: "request.attributes.get('version') >= 22.04"
+        Core\Category\RealTime\Infrastructure\Api\FindHostCategory\FindHostCategoryController
+    condition: "request.attributes.get('version') >= 22.10"
