--- src/Core/Infrastructure/RealTime/Repository/Tag/DbTagFactory.php
@@ -1,39 +1,0 @@
-<?php
-
-/*
- * Copyright 2005 - 2022 Centreon (https://www.centreon.com/)
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- *
- * For more information : contact@centreon.com
- *
- */
-declare(strict_types=1);
-
-namespace Core\Infrastructure\RealTime\Repository\Tag;
-
-use Core\Domain\RealTime\Model\Tag;
-
-class DbTagFactory
-{
-    /**
-     * Create ServiceCategory model using data from database
-     *
-     * @param array<string, mixed> $data
-     * @return Tag
-     */
-    public static function createFromRecord(array $data): Tag
-    {
-        return new Tag((int) $data['id'], $data['name'], (int) $data['type']);
-    }
-}
