+++ src/Core/Tag/RealTime/Application/UseCase/FindTag/FindTagResponse.php
@@ -1,0 +1,44 @@
+<?php
+
+/*
+ * Copyright 2005 - 2022 Centreon (https://www.centreon.com/)
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ * For more information : contact@centreon.com
+ *
+ */
+declare(strict_types=1);
+
+namespace Core\Tag\RealTime\Application\UseCase\FindTag;
+
+use Core\Application\RealTime\Common\RealTimeResponseTrait;
+use Core\Tag\RealTime\Domain\Model\Tag;
+
+class FindTagResponse
+{
+    use RealTimeResponseTrait;
+
+    /**
+     * @var array<int, array<string, int|string>>
+     */
+    public array $tags;
+
+    /**
+     * @param Tag[] $tags
+     */
+    public function __construct(array $tags)
+    {
+        $this->tags = $this->tagsToArray($tags);
+    }
+}
